GJK boolean and distance versions finished
- GJK has been implemented as 2 versions. One with a boolean result that can early out quickly and tell you the answer. The other one which returns the separating distance needs more iterations to find out the answer. Had many issues with the stability of the algorithm and the termination conditions, but most of them have been ironed out (hopefully, no more problems will appear from this)
- Added some mathematics functions, including voronoi simplex region detection, and some plane and triangle related functions
- The sandbox settings inspector now has a new category, the Components header, where actions related to the sandbox components can be performed. At the moment, only resetting all components of a certain type can be done
- Removed some prototyping code

+++ Editor/src/Sandbox/Sandbox.cpp
@@ -232,9 +232,9 @@ static SOLVE_SANDBOX_MODULE_SNAPSHOT_RESULT SolveSandboxModuleSnapshotsChanges(E
 									}
 									else {
 										for (size_t subindex = 0; subindex < snapshot_settings.size && !snapshot_has_changed; subindex++) {
-											size_t existing_setting = FindString(snapshot_settings[subindex].name, module_settings, [](EditorModuleReflectedSetting setting) {
+											unsigned int existing_setting = FindString(snapshot_settings[subindex].name, module_settings, [](EditorModuleReflectedSetting setting) {
 												return setting.name;
-												});
+											});
 											if (existing_setting == -1) {
 												snapshot_has_changed = true;
 											}
@@ -2124,7 +2124,6 @@ bool RenderSandbox(EditorState* editor_state, unsigned int sandbox_index, EDITOR
 
 		// We need to record the query cache to restore it later
 		sandbox->sandbox_world.entity_manager->CopyQueryCache(&runtime_query_cache, &runtime_query_cache_allocator);
-
 		EDITOR_SANDBOX_VIEWPORT active_viewport = GetSandboxActiveViewport(editor_state, sandbox_index);
 		if (active_viewport == EDITOR_SANDBOX_VIEWPORT_SCENE) {
 			viewport_entity_manager = &sandbox->scene_entities;
